code refactored to ES6 and code divided into more files
