+++ package.json
@@ -4,8 +4,11 @@
   "main": "index.js",
   "license": "MIT",
   "dependencies": {
+    "@babel/cli": "^7.4.4",
+    "@babel/core": "^7.4.5",
+    "@babel/preset-env": "^7.4.5",
     "babel-cli": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
+    "babel-env": "^2.4.1",
     "ejs": "^2.6.1",
     "global": "^4.3.2",
     "inquirer": "^6.3.1"
