+++ src/templates/drf/config/settings/base.py
