+++ src/templates/drf/config/settings/local.py
