+++ src/templates/drf/config/settings/test.py
