+++ src/templates/drf/config/urls.py
