+++ src/templates/drf/core/api/apps.py
