+++ src/templates/drf/core/api/urls.ejs
