+++ src/templates/drf/core/api/views.py
