+++ src/templates/drf/deploy/production/nginx.ejs
