+++ src/templates/drf/deploy/staging/nginx.ejs
