+++ src/templates/drf/deploy/staging/supervisor.ejs
