+++ src/templates/drf/fabfile.ejs
