+++ src/templates/drf/manage.py
