+++ src/templates/drf/requirements.txt
