+++ src/templates/drf/requirements/base.txt
