+++ src/templates/drf/travis.ejs
